+++ script/enterprise-server-deprecations/remove-version-markup.js
@@ -26,8 +26,10 @@ program
   .option('-r, --release <NUMBER>', 'Enterprise Server release number. Example: 2.19')
   .parse(process.argv)
 
+const release = program.opts().release
+
 // verify CLI options
-if (!program.release) {
+if (!release) {
   console.log(program.description() + '\n')
   program.options.forEach(opt => {
     console.log(opt.flags)
@@ -36,13 +38,13 @@ if (!program.release) {
   process.exit(1)
 }
 
-if (!enterpriseServerReleases.all.includes(program.release)) {
-  console.log(`You specified ${program.release}! Please specify a supported or deprecated release number from lib/enterprise-server-releases.js`)
+if (!enterpriseServerReleases.all.includes(release)) {
+  console.log(`You specified ${release}! Please specify a supported or deprecated release number from lib/enterprise-server-releases.js`)
   process.exit(1)
 }
 
-const versionToDeprecate = `enterprise-server@${program.release}`
-const currentIndex = indexOf(enterpriseServerReleases.all, program.release)
+const versionToDeprecate = `enterprise-server@${release}`
+const currentIndex = indexOf(enterpriseServerReleases.all, release)
 const nextOldestRelease = nth(enterpriseServerReleases.all, currentIndex - 1)
 const nextOldestVersion = `enterprise-server@${nextOldestRelease}`