+++ translations/zh-CN/content/admin/release-notes.md
@@ -5,6 +5,7 @@ layout: release-notes
 versions:
   enterprise-server: '*'
 topics:
-  - 企业
+  - Enterprise
+  - Upgrades
 ---